Post indexes now show only posts
This was an interesting thing I learned. It seems in future that I should
plan ahead when using STI as the base that everything inherits from doesn't
distinguish type when getting everything. This actually seems like an advantage
if I plan it that way. Adding STI without planning ahead for it seems like
it might cause problems sometimes.